+++ Model/System/Config/Source/HostedpageVersion.php
@@ -1,0 +1,44 @@
+<?php
+/**
+ * HiPay Fullservice Magento
+ *
+ * NOTICE OF LICENSE
+ *
+ * This source file is subject to the Apache 2.0 Licence
+ * that is bundled with this package in the file LICENSE.md.
+ * It is also available through the world-wide-web at this URL:
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * @copyright      Copyright (c) 2016 - HiPay
+ * @license        http://www.apache.org/licenses/LICENSE-2.0 Apache 2.0 Licence
+ *
+ */
+namespace HiPay\FullserviceMagento\Model\System\Config\Source;
+
+/**
+ * Source model for hostedpage versions
+ *
+ * @package HiPay\FullserviceMagento
+ * @author Kassim Belghait <kassim@sirateck.com>
+ * @copyright Copyright (c) 2016 - HiPay
+ * @license http://www.apache.org/licenses/LICENSE-2.0 Apache 2.0 Licence
+ * @link https://github.com/hipay/hipay-fullservice-sdk-magento2
+ */
+class HostedpageVersion implements \Magento\Framework\Option\ArrayInterface
+{
+    const V1 = 'hpv1';
+    const V2 = 'hpv2';
+
+    /**
+     * Options getter
+     *
+     * @return array
+     */
+    public function toOptionArray()
+    {
+        return [
+            ['value' => self::V1, 'label' => __('Hosted Page v1')],
+            ['value' => self::V2, 'label' => __('Hosted Page v2')],
+        ];
+    }
+}