+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - **Add** iDeal Hosted Fields payment method
+- **Add** Hosted Page v2 option on HostedPage parameters
 
 ## 1.12.2