+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - **Add** iDeal Hosted Fields payment method
 - **Add** Hosted Page v2 option on HostedPage parameters
+- **Add** missing fields for Hosted payment methods
 
 ## 1.12.2